refactor: extract SystemEventCellNode base class
CallEventCellNode now inherits from SystemEventCellNode which provides centered layout and label node.
Ready for future system events (membership changes, room updates, etc.).
